Add formatting module.
