+++ metrics-exporter-prometheus/src/lib.rs
@@ -101,7 +101,7 @@
 #![deny(missing_docs)]
 #![cfg_attr(docsrs, feature(doc_cfg), deny(rustdoc::broken_intra_doc_links))]
 mod common;
-pub use self::common::{BuildError, Matcher, sanitize_description, sanitize_label_key, sanitize_label_value, sanitize_metric_name};
+pub use self::common::{BuildError, Matcher};
 
 mod distribution;
 pub use distribution::{Distribution, DistributionBuilder};
@@ -110,4 +110,6 @@ mod builder;
 pub use self::builder::PrometheusBuilder;
 
 mod recorder;
+pub mod formatting;
+
 pub use self::recorder::{PrometheusHandle, PrometheusRecorder};
